Corrected: - made generic DamageWithOnKillEffectAction, including optional argument to control whether it triggers on minion kills (defaults to false) - renamed removeCard metaUtil to DestroyCardPermanently - made Frostbite work on minion kills
