+++ src/main/resources/stsjorbsmodResources/localization/eng/JorbsMod-Card-Strings.json
@@ -197,7 +197,7 @@
   },
   "stsjorbsmod:Frostbite": {
     "NAME": "Frostbite",
-    "DESCRIPTION": "Deal !D! damage. NL If Fatal, stsjorbsmod:Consume this."
+    "DESCRIPTION": "Deal !D! damage. NL If this kills an enemy, stsjorbsmod:Consume this."
   },
   "stsjorbsmod:GatherPower": {
     "NAME": "Gather Power",
